詠ボタンをタイムライン以外にも表示
